+++ SharePointDsc/Examples/Resources/SPUserProfileProperty/1-Example.ps1
@@ -76,6 +76,7 @@ Configuration Example
             IsUserEditable       = $true
             IsAlias              = $false
             IsSearchable         = $false
+            IsReplicable         = $false
             TermStore            = ""
             TermGroup            = ""
             TermSet              = ""